backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==24.1.0
httpx==0.25.0
pillow==10.0.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
pydantic==2.4.2
python-dotenv==1.0.0
modal==0.56.4874